+++ РПГ/Reward.h
@@ -1,0 +1,10 @@
+#pragma once
+
+
+//Класс награды за убийство
+class Reward
+{
+public:
+	int Gold;
+	int Experience;
+};
